fix: use tsx for files with jsx
