+++ .circleci/config.yml
@@ -45,7 +45,7 @@ commands:
   type_check:
     description: 'Running type checkers'
     steps:
-      - run: yarn flow
+      - run: yarn type-check
   check_artefacts_bundlesizes:
     description: 'Running bundlesize check'
     steps:
