+++ packages/react/modules/components/toggle-feature/toggle-feature.ts
@@ -50,7 +50,7 @@ export default class ToggleFeature extends React.PureComponent<Props> {
     children: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
   };
 
-  render(): React.Node {
+  render() {
     if (this.props.isFeatureEnabled) {
       if (this.props.toggledComponent)
         return React.createElement(this.props.toggledComponent);
